+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 ## [2.21.0] 2021-06-02
 
-- hardis:project:clean:references: Improve performances for removing files 
+- hardis:project:clean:references: Improve performances for removing files
 - hardis:scratch:create : Shorten scratch org auto-generated name
 - Authenticate to an org: Request user to set alias if not provided
 - Update default gitlab-ci.yml